Correct minor things in #60:
- Move diagnostics test to separate file
- Parameterize tests for ARIMA rather than loop
- Cover previously ignored lines in test file
- Move safe_mkdirs inside of try/finally
